SpineModel.jl Documentation

SpineModel.jl Documentation

Documentation for SpineModel.jl.

Contents

SpineModel.jl Documentation

Contents

Index

Public interface

Internals

Index

SpineModel.run_spinemodel

SpineModel.run_spinemodel

Public interface

SpineModel.run_spinemodel — Method.

run_spinemodel(
    url_in, url_out;
    optimizer=Cbc.Optimizer,
    cleanup=true,
    extend=m->nothing,
    result=""
)

Run the Spine model from url_in and write results to url_out. At least url_in must point to valid Spine database. A new Spine database is created at url_out if it doesn't exist.

Optional keyword arguments

optimizer is the constructor of the optimizer used for building and solving the model.

cleanup tells run_spinemodel whether or not convenience function callables should be set to nothing after completion.

extend is a function for extending the model. run_spinemodel calls this function with the internal JuMP.Model object before calling JuMP.optimize!.

result is the name of the result object to write to url_out when saving results.

source

SpineModel.run_spinemodel — Method.

run_spinemodel(
    url;
    optimizer=Cbc.Optimizer,
    cleanup=true,
    extend=m->nothing,
    result=""
)

Run the Spine model from url and write results to the same url. Keyword arguments have the same purpose as for run_spinemodel(::String, ::String).

source

Internals

TODO