Library

Library

Documentation for SpineModel.jl.

Contents

Library

Contents

Index

Public interface

Internals

Variables

Constraints

Objectives

Index

SpineModel.fix_flow_indices

SpineModel.fix_trans_indices

SpineModel.fix_units_on_indices

SpineModel.flow_indices

SpineModel.run_spinemodel

SpineModel.run_spinemodel

SpineModel.trans_indices

SpineModel.units_on_indices

SpineModel.var_flow_indices

SpineModel.var_trans_indices

SpineModel.var_units_on_indices

SpineModel.variable_flow

SpineModel.variable_trans

SpineModel.variable_units_on

Public interface

SpineModel.run_spinemodel — Method.

run_spinemodel(
    url_in, url_out;
    optimizer=Cbc.Optimizer,
    cleanup=true,
    extend=m->nothing,
    result=""
)

Run the Spine model from url_in and write results to url_out. At least url_in must point to valid Spine database. A new Spine database is created at url_out if it doesn't exist.

Optional keyword arguments

optimizer is the constructor of the optimizer used for building and solving the model.

cleanup tells run_spinemodel whether or not convenience function callables should be set to nothing after completion.

extend is a function for extending the model. run_spinemodel calls this function with the internal JuMP.Model object before calling JuMP.optimize!.

result is the name of the result object to write to url_out when saving results. An empty string (the default) gets replaced by "result" with the current time appended.

source

SpineModel.run_spinemodel — Method.

run_spinemodel(
    url;
    optimizer=Cbc.Optimizer,
    cleanup=true,
    extend=m->nothing,
    result=""
)

Run the Spine model from url and write results to the same url. Keyword arguments have the same purpose as for run_spinemodel.

source

Internals

Variables

SpineModel.variable_flow — Function.

variable_flow(m::Model)

Create the flow variable for the model m.

This variable represents the (average) instantaneous flow of a commodity between a node and a unit in a certain direction and within a certain time slice.

source

SpineModel.variable_trans — Function.

variable_trans(m::Model)

Create the trans variable for model m.

This variable represents the (average) instantaneous flow of a commodity between a node and a connection in a certain direction and within a certain time slice.

source

SpineModel.variable_units_on — Function.

variable_units_on(m::Model)

Create the units_on variable for model m.

This variable represents the number of online units for a given unit within a certain time slice.

source

SpineModel.flow_indices — Function.

flow_indices(
    commodity=anything,
    node=anything,
    unit=anything,
    direction=anything,
    t=anything
)

A list of NamedTuples corresponding to indices of the flow variable. The keyword arguments act as filters for each dimension.

source

SpineModel.var_flow_indices — Function.

var_flow_indices(
    commodity=anything,
    node=anything,
    unit=anything,
    direction=anything,
    t=anything
)

A list of NamedTuples corresponding to non-fixed indices of the flow variable. The keyword arguments act as filters for each dimension.

source

SpineModel.fix_flow_indices — Function.

fix_flow_indices(
    commodity=anything,
    node=anything,
    unit=anything,
    direction=anything,
    t=anything
)

A list of NamedTuples corresponding to fixed indices of the flow variable. The keyword arguments act as filters for each dimension.

source

SpineModel.trans_indices — Function.

trans_indices(
    commodity=anything,
    node=anything,
    connection=anything,
    direction=anything,
    t=anything
)

A list of NamedTuples corresponding to indices of the trans variable. The keyword arguments act as filters for each dimension.

source

SpineModel.var_trans_indices — Function.

var_trans_indices(
    commodity=anything,
    node=anything,
    connection=anything,
    direction=anything,
    t=anything
)

A list of NamedTuples corresponding to non-fixed indices of the trans variable. The keyword arguments act as filters for each dimension.

source

SpineModel.fix_trans_indices — Function.

fix_trans_indices(
    commodity=anything,
    node=anything,
    connection=anything,
    direction=anything,
    t=anything
)

A list of NamedTuples corresponding to fixed indices of the trans variable. The keyword arguments act as filters for each dimension.

source

SpineModel.units_on_indices — Function.

units_on_indices(unit=anything, t=anything)

A list of NamedTuples corresponding to indices of the units_on variable. The keyword arguments act as filters for each dimension.

source

SpineModel.var_units_on_indices — Function.

var_units_on_indices(unit=anything, t=anything)

A list of NamedTuples corresponding to non_fixed indices of the units_on variable. The keyword arguments act as filters for each dimension.

source

SpineModel.fix_units_on_indices — Function.

fix_units_on_indices(unit=anything, t=anything)

A list of NamedTuples corresponding to fixed indices of the units_on variable. The keyword arguments act as filters for each dimension.

source

Constraints

TODO

Objectives

TODO