Library

Library

Documentation for SpineModel.jl.

Contents

Library

Contents

Index

Public interface

Internals

Variables

Constraints

Objectives

Index

SpineModel.flow_indices

SpineModel.run_spinemodel

SpineModel.run_spinemodel

SpineModel.trans_indices

SpineModel.units_on_indices

Public interface

SpineModel.run_spinemodel — Method.

run_spinemodel(url_in, url_out; <keyword arguments>)

Run the Spine model from url_in and write report to url_out. At least url_in must point to valid Spine database. A new Spine database is created at url_out if it doesn't exist.

Keyword arguments

with_optimizer=with_optimizer(Cbc.Optimizer, logLevel=0) is the optimizer factory for building the JuMP model.

cleanup=true tells run_spinemodel whether or not convenience functors should be set to nothing after completion.

add_constraints=m -> nothing is called with the Model object in the first optimization window, and allows adding user contraints.

update_constraints=m -> nothing is called in windows 2 to the last, and allows updating contraints added by add_constraints.

log_level=3 is the log level.

source

SpineModel.run_spinemodel — Method.

run_spinemodel(url; <keyword arguments>)

Run the Spine model from url and write report to the same url. Keyword arguments have the same purpose as for run_spinemodel.

source

Internals

Variables

Missing docstring.

Missing docstring for variable_flow. Check Documenter's build log for details.

Missing docstring.

Missing docstring for variable_trans. Check Documenter's build log for details.

Missing docstring.

Missing docstring for variable_units_on. Check Documenter's build log for details.

SpineModel.flow_indices — Function.

flow_indices(
    commodity=anything,
    node=anything,
    unit=anything,
    direction=anything,
    t=anything
)

A list of NamedTuples corresponding to indices of the flow variable. The keyword arguments act as filters for each dimension.

source

Missing docstring.

Missing docstring for var_flow_indices. Check Documenter's build log for details.

Missing docstring.

Missing docstring for fix_flow_indices. Check Documenter's build log for details.

SpineModel.trans_indices — Function.

trans_indices(
    commodity=anything,
    node=anything,
    connection=anything,
    direction=anything,
    t=anything
)

A list of NamedTuples corresponding to indices of the trans variable. The keyword arguments act as filters for each dimension.

source

Missing docstring.

Missing docstring for var_trans_indices. Check Documenter's build log for details.

Missing docstring.

Missing docstring for fix_trans_indices. Check Documenter's build log for details.

SpineModel.units_on_indices — Function.

units_on_indices(unit=anything, t=anything)

A list of NamedTuples corresponding to indices of the units_on variable. The keyword arguments act as filters for each dimension.

source

Missing docstring.

Missing docstring for var_units_on_indices. Check Documenter's build log for details.

Missing docstring.

Missing docstring for fix_units_on_indices. Check Documenter's build log for details.

Constraints

TODO

Objectives

TODO