Library

Documentation for SpineOpt.jl.

Contents

Library

Contents

Index

Public interface

Internals

Variable library

Constraint library

Objective

Index

SpineOpt.run_spineopt

Public interface

SpineOpt.run_spineopt — Function

run_spineopt(url_in, url_out; <keyword arguments>)

Run the SpineOpt from url_in and write report to url_out. At least url_in must point to valid Spine database. A new Spine database is created at url_out if it doesn't exist.

Keyword arguments

with_optimizer=with_optimizer(Cbc.Optimizer, logLevel=0) is the optimizer factory for building the JuMP model.

cleanup=true tells run_spineopt whether or not convenience functors should be set to nothing after completion.

add_constraints=m -> nothing is called with the Model object in the first optimization window, and allows adding user contraints.

update_constraints=m -> nothing is called in windows 2 to the last, and allows updating contraints added by add_constraints.

log_level=3 is the log level.

TODO

Internals

Variable library

Missing docstring.

Missing docstring for unit_flow_indices. Check Documenter's build log for details.

TODO

Constraint library

TODO

Objective

TODO