Library

Documentation for SpineOpt.jl.

Contents

Library

Contents

Index

Public interface

Internals

Variable library

Constraint library

Objective

Index

SpineOpt.active_stochastic_paths

SpineOpt.generate_forced_availability_factor

SpineOpt.run_spineopt

SpineOpt.t_before_t

SpineOpt.t_in_t

SpineOpt.t_overlaps_t

SpineOpt.time_slice

SpineOpt.unit_flow_indices

SpineOpt.write_report_from_intermediate_results

Public interface

SpineOpt.run_spineopt — Function

run_spineopt(url_in, url_out; <keyword arguments>)

Run SpineOpt using the contents of url_in and write report(s) to url_out. At least url_in must point to a valid Spine database. A new Spine database is created at url_out if one doesn't exist.

Arguments

upgrade::Bool=false: whether or not to automatically upgrade the data structure in url_in to latest.

mip_solver=nothing: a MIP solver to use if no MIP solver specified in the DB.

lp_solver=nothing: a LP solver to use if no LP solver specified in the DB.

add_constraints=m -> nothing: a function that receives the Model object as argument and adds custom user constraints.

log_level::Int=3: an integer to control the log level.

optimize::Bool=true: whether or not to optimise the model (useful for running tests).

update_names::Bool=false: whether or not to update variable and constraint names after the model rolls (expensive).

alternative::String="": if non empty, write results to the given alternative in the output DB.

write_as_roll::Int=0: if greater than 0 and the run has a rolling horizon, then write results every that many windows.

use_direct_model::Bool=false: whether or not to use JuMP.direct_model to build the Model object.

filters::Dict{String,String}=Dict("tool" => "object_activity_control"): a dictionary to specify filters. Possible keys are "tool" and "scenario". Values should be a tool or scenario name in the input DB.

log_file_path::String=nothing: if not nothing, log all console output to a file at the given path. The file is overwritten at each call.

resume_file_path::String=nothing: only relevant in rolling horizon optimisations with write_as_roll greater or equal than one. If the file at given path contains resume data from a previous run, start the run from that point. Also, save resume data to that same file as the model rolls and results are written to the output database.

Example

using SpineOpt
m = run_spineopt(
    raw"sqlite:///C:\path\to\your\inputputdb.sqlite", 
    raw"sqlite:///C:\path\to\your\outputdb.sqlite";
    filters=Dict("tool" => "object_activity_control", "scenario" => "scenario_to_run"),
    alternative="your_results_alternative"
)

SpineOpt.write_report_from_intermediate_results — Function

write_report_from_intermediate_results(intermediate_results_folder, default_url; <keyword arguments>)

Collect results generated on a previous, unsuccessful SpineOpt run from intermediate_results_folder, and write the corresponding report(s) to url_out. A new Spine database is created at url_out if one doesn't exist.

Arguments

alternative::String="": if non empty, write results to the given alternative in the output DB.

log_level::Int=3: an integer to control the log level.

SpineOpt.generate_forced_availability_factor — Function

generate_forced_availability_factor(url_in, url_out; <keyword arguments>)

Generate forced availability factors (due to outages) from the contents of url_in and write them to url_out. At least url_in must point to a valid Spine database. A new Spine database is created at url_out if one doesn't exist.

To generate forced availability factors for an entity, specify mean_time_to_failure and optionally mean_time_to_repair for that entity as a duration in the input DB.

Parameter forced_availability_factor will be written for those entites in the output DB, holding a time series with the availability factor due to forced outages.

Arguments

alternative::String="": if non empty, write results to the given alternative in the output DB.

filters::Dict{String,String}=Dict("tool" => "object_activity_control"): a dictionary to specify filters. Possible keys are "tool" and "scenario". Values should be a tool or scenario name in the input DB.

Example

using SpineOpt
m = generate_forced_availability_factor(
    raw"sqlite:///C:\path\to\your\inputputdb.sqlite", 
    raw"sqlite:///C:\path\to\your\outputdb.sqlite"
)

SpineOpt.time_slice — Function

time_slice(m; temporal_block=anything, t=anything)

An Array of TimeSlices in model m.

Keyword arguments

temporal_block: only return TimeSlices in this block or blocks.

t: only return time slices from this collection.

SpineOpt.t_in_t — Function

t_in_t(m; t_short=anything, t_long=anything)

An Array where each element is a Tuple of two TimeSlices in model m, the second containing the first.

Keyword arguments

t_short: if given, return an Array of TimeSlices that contain t_short.

t_long: if given, return an Array of TimeSlices that are contained in t_long.

SpineOpt.t_before_t — Function

t_before_t(m; t_before=anything, t_after=anything)

An Array where each element is a Tuple of two consecutive TimeSlices in model m (the second starting when the first ends).

Keyword arguments

t_before: if given, return an Array of TimeSlices that start when t_before ends.

t_after: if given, return an Array of TimeSlices that end when t_after starts.

SpineOpt.t_overlaps_t — Function

t_overlaps_t(m; t)

An Array of TimeSlices in model m that overlap the given t.

SpineOpt.active_stochastic_paths — Function

active_stochastic_paths(m; stochastic_structure, t)

An Array where each element is itself an Array of stochastic_scenario Objects, corresponding to a branch (or path) of the stochastic DAG associated to model m.

This DAG has one vertex per stochastic_scenario, and the edges are given by the parent_stochastic_scenario__child_stochastic_scenario relationships.

The result is obtained by first taking the subset of the stochastic DAG defined by stochastic_structure (which can be a single Object or an Array of Objects); and then taking the branches of that subset that cover t (which can be a single TimeSlice or an Array of TimeSlices)

TODO

Internals

Variable library

SpineOpt.unit_flow_indices — Function

unit_flow_indices(
    unit=anything,
    node=anything,
    direction=anything,
    s=anything
    t=anything
)

A list of NamedTuples corresponding to indices of the unit_flow variable where the keyword arguments act as filters for each dimension.

TODO

Constraint library

TODO

Objective

TODO