Library

Documentation for SpineOpt.jl.

Contents

Library

Contents

Index

Public interface

Internals

Variables

Constraints

Objectives

Index

SpineOpt.connection_flow_indices

SpineOpt.run_spineopt

SpineOpt.run_spineopt

SpineOpt.units_on_indices

Public interface

SpineOpt.run_spineopt — Method

run_spineopt(url_in, [url_out=url_in]; <keyword arguments>)

Run SpineOpt from url_in and write report to url_out. A new Spine database is created at url_out if it doesn't exist.

Keyword arguments

with_optimizer=with_optimizer(Cbc.Optimizer, logLevel=0) is the optimizer factory for building the JuMP model.

cleanup=true tells run_spineopt whether or not convenience functors should be set to nothing after completion.

add_constraints=m -> nothing is called with the Model object in the first optimization window, and allows adding user contraints.

update_constraints=m -> nothing is called in windows 2 to the last, and allows updating contraints added by add_constraints.

log_level=3 is the log level.

SpineOpt.run_spineopt — Method

run_spineopt(url_in, [url_out=url_in]; <keyword arguments>)

Run SpineOpt from url_in and write report to url_out. A new Spine database is created at url_out if it doesn't exist.

Keyword arguments

with_optimizer=with_optimizer(Cbc.Optimizer, logLevel=0) is the optimizer factory for building the JuMP model.

cleanup=true tells run_spineopt whether or not convenience functors should be set to nothing after completion.

add_constraints=m -> nothing is called with the Model object in the first optimization window, and allows adding user contraints.

update_constraints=m -> nothing is called in windows 2 to the last, and allows updating contraints added by add_constraints.

log_level=3 is the log level.

Internals

Variables

Missing docstring.

Missing docstring for variable_flow. Check Documenter's build log for details.

Missing docstring.

Missing docstring for variable_connection_flow. Check Documenter's build log for details.

Missing docstring.

Missing docstring for variable_units_on. Check Documenter's build log for details.

Missing docstring.

Missing docstring for flow_indices. Check Documenter's build log for details.

Missing docstring.

Missing docstring for var_flow_indices. Check Documenter's build log for details.

Missing docstring.

Missing docstring for fix_unit_flow_indices. Check Documenter's build log for details.

SpineOpt.connection_flow_indices — Function

connection_flow_indices(
    connection=anything,
    node=anything,
    direction=anything,
    t=anything
)

A list of NamedTuples corresponding to indices of the connection_flow variable. The keyword arguments act as filters for each dimension.

Missing docstring.

Missing docstring for var_connection_flow_indices. Check Documenter's build log for details.

Missing docstring.

Missing docstring for fix_connection_flow_indices. Check Documenter's build log for details.

SpineOpt.units_on_indices — Function

units_on_indices(unit=anything, stochastic_scenario=anything, t=anything)

A list of NamedTuples corresponding to indices of the units_on variable. The keyword arguments act as filters for each dimension.

Missing docstring.

Missing docstring for var_units_on_indices. Check Documenter's build log for details.

Missing docstring.

Missing docstring for fix_units_on_indices. Check Documenter's build log for details.

Constraints

TODO

Objectives

TODO